Export SleepActionReceiver to remotely act on the app
Closes #174
